+++ parallel-install/pkg/deployment/core.go
@@ -2,14 +2,8 @@
 package deployment
 
 import (
-	"context"
-	"strings"
 	"time"
-	"unicode"
 
-	"k8s.io/apimachinery/pkg/labels"
-
-	"github.com/avast/retry-go"
 	"github.com/pkg/errors"
 
 	"github.com/kyma-incubator/hydroform/parallel-install/pkg/components"
@@ -18,14 +12,13 @@ import (
 	"github.com/kyma-incubator/hydroform/parallel-install/pkg/helm"
 	"github.com/kyma-incubator/hydroform/parallel-install/pkg/logger"
 	"github.com/kyma-incubator/hydroform/parallel-install/pkg/overrides"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 )
 
 type core struct {
 	// Contains list of components to install (inclusive pre-requisites)
 	cfg       *config.Config
-	overrides *OverridesBuilder
+	overrides *overrides.Builder
 	// Used to send progress events of a running install/uninstall process
 	processUpdates func(ProcessUpdate)
 	kubeClient     kubernetes.Interface
@@ -40,7 +33,7 @@ type core struct {
 //kubeClient is the kubernetes client
 //
 //processUpdates can be an optional feedback channel provided by the caller
-func newCore(cfg *config.Config, overrides *OverridesBuilder, kubeClient kubernetes.Interface, processUpdates func(ProcessUpdate)) *core {
+func newCore(cfg *config.Config, overrides *overrides.Builder, kubeClient kubernetes.Interface, processUpdates func(ProcessUpdate)) *core {
 	return &core{
 		cfg:            cfg,
 		overrides:      overrides,
@@ -126,85 +119,14 @@ func (i *core) processUpdateComponent(phase InstallationPhase, comp components.K
 	})
 }
 
-func isK3dCluster(kubeClient kubernetes.Interface) (isK3d bool, err error) {
-
-	retryOptions := []retry.Option{
-		retry.Delay(2 * time.Second),
-		retry.Attempts(3),
-		retry.DelayType(retry.FixedDelay),
-	}
-
-	err = retry.Do(func() error {
-		nodeList, err := kubeClient.CoreV1().Nodes().List(context.Background(), metav1.ListOptions{})
-		if err != nil {
-			return err
-		}
-
-		for _, node := range nodeList.Items {
-			if strings.HasPrefix(node.GetName(), "k3d-") {
-				isK3d = true
-				return nil
-			}
-		}
-
-		return nil
-	}, retryOptions...)
-	if err != nil {
-		return isK3d, err
-	}
-
-	return isK3d, nil
-}
-
-func getK3dClusterName(kubeClient kubernetes.Interface) (k3dName string, err error) {
-	retryOptions := []retry.Option{
-		retry.Delay(2 * time.Second),
-		retry.Attempts(3),
-		retry.DelayType(retry.FixedDelay),
-	}
-
-	err = retry.Do(func() error {
-		labelSelector := metav1.LabelSelector{
-			MatchLabels: map[string]string{"node-role.kubernetes.io/master": "true"},
-		}
-		listOptions := metav1.ListOptions{
-			LabelSelector: labels.Set(labelSelector.MatchLabels).String(),
-		}
-		nodeList, err := kubeClient.CoreV1().Nodes().List(context.Background(), listOptions)
-		if err != nil {
-			return err
-		}
-
-		for _, node := range nodeList.Items {
-			nodeName := node.GetName()
-			if !strings.HasPrefix(nodeName, "k3d-") {
-				k3dName = ""
-				return errors.New("Cluster is not a k3d cluster")
-			}
-			// K3d cluster name can be derived from master node names, which has the form k3d-<cluster-name>-server-<id>.
-			// E.g., with the Kyma CLI default flags k3d-kyma-server-0
-			k3dName = strings.TrimSuffix(strings.TrimRightFunc(strings.TrimPrefix(nodeName, "k3d-"), func(r rune) bool {
-				return unicode.IsNumber(r) || r == '-'
-			}), "-server")
-		}
-
-		return nil
-	}, retryOptions...)
-	if err != nil {
-		return k3dName, err
-	}
-
-	return k3dName, nil
-}
-
-func registerOverridesInterceptors(ob *OverridesBuilder, kubeClient kubernetes.Interface, log logger.Interface) {
+func registerOverridesInterceptors(ob *overrides.Builder, kubeClient kubernetes.Interface, log logger.Interface) {
 	//hide certificate data
-	ob.AddInterceptor([]string{"global.domainName", "global.ingress.domainName"}, NewDomainNameOverrideInterceptor(kubeClient, log))
-	ob.AddInterceptor([]string{"global.tlsCrt", "global.tlsKey"}, NewCertificateOverrideInterceptor("global.tlsCrt", "global.tlsKey", kubeClient))
+	ob.AddInterceptor([]string{"global.domainName", "global.ingress.domainName"}, overrides.NewDomainNameOverrideInterceptor(kubeClient, log))
+	ob.AddInterceptor([]string{"global.tlsCrt", "global.tlsKey"}, overrides.NewCertificateOverrideInterceptor("global.tlsCrt", "global.tlsKey", kubeClient))
 
 	// make sure k3d clusters use k3d container registry
-	ob.AddInterceptor([]string{"serverless.dockerRegistry.internalServerAddress", "serverless.dockerRegistry.serverAddress", "serverless.dockerRegistry.registryAddress"}, NewRegistryInterceptor(kubeClient))
+	ob.AddInterceptor([]string{"serverless.dockerRegistry.internalServerAddress", "serverless.dockerRegistry.serverAddress", "serverless.dockerRegistry.registryAddress"}, overrides.NewRegistryInterceptor(kubeClient))
 
 	// make sure k3d clusters disable internal container registry
-	ob.AddInterceptor([]string{"serverless.dockerRegistry.enableInternal"}, NewRegistryDisableInterceptor(kubeClient))
+	ob.AddInterceptor([]string{"serverless.dockerRegistry.enableInternal"}, overrides.NewRegistryDisableInterceptor(kubeClient))
 }
