+++ parallel-install/pkg/overrides/overrides_test.go
@@ -4,131 +4,64 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/kyma-incubator/hydroform/parallel-install/pkg/logger"
-
 	"github.com/stretchr/testify/require"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes/fake"
-
 	"gopkg.in/yaml.v3"
 )
 
-func Test_ReadOverridesFromCluster(t *testing.T) {
-	component := "monitoring"
-
-	// fake k8s with override ConfigMaps
-	k8sMock := fake.NewSimpleClientset(
-		&v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "monitoring-overrides",
-				Namespace: "kyma-installer",
-				Labels:    map[string]string{"installer": "overrides", "component": component},
-			},
-			Data: map[string]string{
-				"componentOverride1": "test1",
-				"componentOverride2": "test2",
-			},
-		},
-		&v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "global-overrides",
-				Namespace: "kyma-installer",
-				Labels:    map[string]string{"installer": "overrides"},
-			},
-			Data: map[string]string{
-				"global.globalOverride1": "test1",
-				"global.globalOverride2": "test2",
-			},
-		},
-	)
-
-	// Read additional overrides file
-	data, err := ioutil.ReadFile("../test/data/overrides.yaml")
+func Test_MergeOverrides(t *testing.T) {
+	builder := Builder{}
+	err := builder.AddFile("../test/data/deployment-overrides1.yaml")
 	require.NoError(t, err)
-	var overrides map[string]interface{}
-	err = yaml.Unmarshal(data, &overrides)
+	err = builder.AddFile("../test/data/deployment-overrides2.json")
 	require.NoError(t, err)
 
-	data, err = ioutil.ReadFile("../test/data/overrides-colliding.yaml")
+	override1 := make(map[string]interface{})
+	override1["key4"] = "value4override1"
+	err = builder.AddOverrides("chart", override1)
 	require.NoError(t, err)
-	var overridesColliding map[string]interface{}
-	err = yaml.Unmarshal(data, &overridesColliding)
+
+	override2 := make(map[string]interface{})
+	override2["key5"] = "value5override2"
+	err = builder.AddOverrides("chart", override2)
 	require.NoError(t, err)
-	log := logger.NewLogger(true)
 
-	t.Run("Should properly read overrides with no colliding data", func(t *testing.T) {
-		testProvider, err := New(k8sMock, overrides, log)
-		require.NoError(t, err)
+	// read expected result
+	data, err := ioutil.ReadFile("../test/data/deployment-overrides-result.yaml")
+	require.NoError(t, err)
+	var expected map[string]interface{}
+	err = yaml.Unmarshal(data, &expected)
+	require.NoError(t, err)
 
-		err = testProvider.ReadOverridesFromCluster()
-		require.NoError(t, err)
+	// verify merge result with expected data
+	result, err := builder.Build()
+	require.NoError(t, err)
+	require.Equal(t, expected, result.Map())
+}
 
-		// Monitoring should have two sample component overrides + one from additional overrides
-		// + global overrides under one "global" key
-		res := testProvider.OverridesGetterFunctionFor(component)()
-		require.Equal(t, 4, len(res), "Number of component overrides not as expected")
+func Test_AddFile(t *testing.T) {
+	builder := Builder{}
+	err := builder.AddFile("../test/data/deployment-overrides1.yaml")
+	require.NoError(t, err)
+	err = builder.AddFile("../test/data/deployment-overrides2.json")
+	require.NoError(t, err)
+	err = builder.AddFile("../test/data/overrides.xml") // unsupported format
+	require.Error(t, err)
+}
 
-		// Another component without any component override should only have two global overrides + one from additional overrides
-		res2 := testProvider.OverridesGetterFunctionFor("anotherComponent")()
-		require.Contains(t, res2, "global")
-		require.Equal(t, 1, len(res2))
-		require.Equal(t, 3, len(res2["global"].(map[string]interface{})), "Number of global overrides not as expected")
-	})
+func Test_AddOverrides(t *testing.T) {
+	builder := Builder{}
+	data := make(map[string]interface{})
 
-	t.Run("Should not duplicate additional overrides when reading overrides many times", func(t *testing.T) {
-		testProvider, err := New(k8sMock, overrides, log)
-		require.NoError(t, err)
+	// invalid
+	err := builder.AddOverrides("", data)
+	require.Error(t, err)
 
-		err = testProvider.ReadOverridesFromCluster()
-		require.NoError(t, err)
+	//invalid
+	err = builder.AddOverrides("xyz", data)
+	require.Error(t, err)
 
-		err = testProvider.ReadOverridesFromCluster()
-		require.NoError(t, err)
-
-		err = testProvider.ReadOverridesFromCluster()
-		require.NoError(t, err)
-
-		// Monitoring should have two sample component overrides + one from additional overrides
-		// + global overrides under one "global" key
-		res := testProvider.OverridesGetterFunctionFor(component)()
-		require.Equal(t, 4, len(res), "Number of component overrides not as expected")
-
-		// Another component without any component override should only have two global overrides + one from additional overrides
-		res2 := testProvider.OverridesGetterFunctionFor("anotherComponent")()
-		require.Contains(t, res2, "global")
-		require.Equal(t, 1, len(res2))
-		require.Equal(t, 3, len(res2["global"].(map[string]interface{})), "Number of global overrides not as expected")
-	})
-
-	t.Run("Should always put additionalOverrides on top of other overrides", func(t *testing.T) {
-		testProvider, err := New(k8sMock, overridesColliding, log)
-		require.NoError(t, err)
-
-		err = testProvider.ReadOverridesFromCluster()
-		require.NoError(t, err)
-
-		// Monitoring should have two sample component overrides with one overridden by additional overrides
-		// + global overrides under one "global" key - two global overrides from this file with one overridden by additional overrides
-		res := testProvider.OverridesGetterFunctionFor(component)()
-		require.Equal(t, 3, len(res), "Number of component overrides not as expected")
-		require.Equal(t, "changed", res["componentOverride1"], "Override from additional overrides not on top of regular overrides")
-
-		// Another component without any component override should only have two global overrides with one overridden by additional overrides
-		res2 := testProvider.OverridesGetterFunctionFor("anotherComponent")()
-		require.Contains(t, res2, "global")
-		require.Equal(t, 1, len(res2))
-		globalOverrides, ok := res2["global"].(map[string]interface{})
-		require.True(t, ok)
-		require.Equal(t, 2, len(res2["global"].(map[string]interface{})), "Number of global overrides not as expected")
-		require.Equal(t, "changed", globalOverrides["globalOverride1"], "Override from additional overrides not on top of regular overrides")
-	})
-
-	t.Run("Test non-nested overrides", func(t *testing.T) {
-		illegalOverrides := make(map[string]interface{})
-		illegalOverrides["componentX"] = "value1"
-		illegalOverrides["componentY"] = "value2"
-		_, err := New(k8sMock, illegalOverrides, log)
-		require.Error(t, err)
-	})
+	//valid
+	data["test"] = "abc"
+	err = builder.AddOverrides("xyz", data)
+	require.NoError(t, err)
 }
